load local changes for PRs and check runs
