+++ controller/integrations/events/controller.go
@@ -2,28 +2,20 @@ package events
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"net/url"
 
-	"github.com/bradleyfalzon/ghinstallation"
 	"github.com/google/go-github/v43/github"
-	"github.com/opencontainers/go-digest"
-	"github.com/vito/bass-loop/pkg/bassgh"
 	"github.com/vito/bass-loop/pkg/blobs"
 	"github.com/vito/bass-loop/pkg/cfg"
 	"github.com/vito/bass-loop/pkg/ghapp"
 	"github.com/vito/bass-loop/pkg/logs"
 	"github.com/vito/bass-loop/pkg/models"
-	"github.com/vito/bass-loop/pkg/runs"
 	"github.com/vito/bass/pkg/bass"
-	"github.com/vito/bass/pkg/cli"
-	"github.com/vito/bass/pkg/ioctx"
 	"github.com/vito/bass/pkg/proto"
 	"github.com/vito/bass/pkg/runtimes"
 	"github.com/vito/bass/pkg/zapctx"
-	"github.com/vito/progrock"
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/grpc"
@@ -42,6 +34,7 @@ type Controller struct {
 }
 
 const DefaultExternalURL = "http://localhost:3000"
+const ProjectFile = "project.bass"
 
 func Load(log *logs.Logger, config *cfg.Config, db *models.Conn, blobs *blobs.Bucket, transport *ghapp.Transport) *Controller {
 	e := config.ExternalURL
@@ -70,216 +63,8 @@ func Load(log *logs.Logger, config *cfg.Config, db *models.Conn, blobs *blobs.Bu
 func (c *Controller) Create(w http.ResponseWriter, r *http.Request) {
 	switch r.URL.Query().Get("integration_id") {
 	case "github":
-		c.handleGithub(w, r)
+		c.handleGitHub(w, r)
 	}
-}
-
-func (c *Controller) handleGithub(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-
-	eventName := r.Header.Get("X-GitHub-Event")
-	deliveryID := r.Header.Get("X-GitHub-Delivery")
-
-	if eventName == "" {
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprintln(w, "missing event type")
-		return
-	}
-
-	payloadBytes, err := github.ValidatePayload(r, []byte(c.Config.GitHubApp.WebhookSecret))
-	if err != nil {
-		w.WriteHeader(http.StatusUnprocessableEntity)
-		fmt.Fprintln(w, "missing event type")
-		return
-	}
-
-	err = c.Handle(ctx, eventName, deliveryID, payloadBytes)
-	if err != nil {
-		cli.WriteError(ctx, err)
-
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprintln(w, err.Error())
-		return
-	}
-}
-
-func (c *Controller) Handle(ctx context.Context, eventName, deliveryID string, payload []byte) error {
-	logger := zapctx.FromContext(ctx).With(
-		zap.String("event", eventName),
-		zap.String("delivery", deliveryID),
-	)
-
-	logger.Info("handling")
-
-	var payloadScope *bass.Scope
-	err := json.Unmarshal(payload, &payloadScope)
-	if err != nil {
-		return fmt.Errorf("payload->scope: %w", err)
-	}
-
-	// attempt to parse a check suite event from the payload so that we can use
-	// its sha instead of the default branch for faster iteration on check
-	// changes
-	//
-	// if it's not a CheckSuiteEvent we'll at least use the Repo field to route
-	// the event
-	var maybeSuiteEvent github.CheckSuiteEvent
-	err = json.Unmarshal(payload, &maybeSuiteEvent)
-	if err != nil {
-		return fmt.Errorf("unmarshal check suite: %w", err)
-	}
-
-	if maybeSuiteEvent.Sender == nil || maybeSuiteEvent.Repo == nil || maybeSuiteEvent.Installation == nil {
-		// be defensive just because we don't really know what events we'll receive
-		logger.Warn("ignoring unknown event")
-		return nil
-	}
-
-	logger = logger.With(
-		zap.String("sender", maybeSuiteEvent.Sender.GetLogin()),
-		zap.String("repo", maybeSuiteEvent.Repo.GetFullName()),
-	)
-
-	// handle the rest async
-	c.dispatches.Go(func() error {
-		defer func() {
-			// we're forking a goroutine from a goroutine, so prevent panics from
-			// taking down the whole loop
-			if err := recover(); err != nil {
-				logger.Error("dispatch panic!", zap.Any("recovered", err))
-			}
-		}()
-
-		err := c.dispatch(
-			zapctx.ToContext(context.Background(), logger),
-			maybeSuiteEvent,
-			eventName,
-			deliveryID,
-			payloadScope,
-		)
-		if err != nil {
-			logger.Warn("dispatch errored", zap.Error(err))
-		}
-
-		return nil
-	})
-
-	return nil
-}
-
-func (c *Controller) dispatch(ctx context.Context, maybeSuiteEvent github.CheckSuiteEvent, eventName, deliveryID string, payloadScope *bass.Scope) error {
-	// each concurrent Bass must have its own trace
-	ctx = bass.WithTrace(context.Background(), &bass.Trace{})
-
-	instID := maybeSuiteEvent.GetInstallation().GetID()
-	sender := maybeSuiteEvent.GetSender()
-	repo := maybeSuiteEvent.GetRepo()
-
-	// load the user's forwarded runtime pool
-	ctx, pool, err := c.withUserPool(ctx, sender)
-	if err != nil {
-		return fmt.Errorf("user %s (%s) pool: %w", sender.GetLogin(), sender.GetNodeID(), err)
-	}
-	defer pool.Close()
-
-	ghClient := github.NewClient(&http.Client{
-		Transport: ghinstallation.NewFromAppsTransport(c.Transport, instID),
-	})
-
-	var ref string
-	if maybeSuiteEvent.CheckSuite != nil {
-		ref = maybeSuiteEvent.CheckSuite.GetHeadSHA()
-	} else {
-		branch, _, err := ghClient.Repositories.GetBranch(ctx, repo.GetOwner().GetLogin(), repo.GetName(), repo.GetDefaultBranch(), true)
-		if err != nil {
-			return fmt.Errorf("get branch: %w", err)
-		}
-
-		ref = branch.GetCommit().GetSHA()
-	}
-
-	hookThunk := bass.Thunk{
-		Cmd: bass.ThunkCmd{
-			FS: bassgh.NewFS(ctx, ghClient, repo, ref, "project.bass"),
-		},
-	}
-
-	run, err := models.CreateThunkRun(ctx, c.DB, sender, hookThunk)
-	if err != nil {
-		return fmt.Errorf("create hook thunk run: %w", err)
-	}
-
-	progress := cli.NewProgress()
-
-	recorder := progrock.NewRecorder(progress)
-	thunkCtx := progrock.RecorderToContext(ctx, recorder)
-
-	rec := recorder.Vertex(digest.Digest("delivery:"+deliveryID), fmt.Sprintf("[delivery] %s %s", eventName, deliveryID))
-	logger := bass.LoggerTo(rec.Stderr())
-	thunkCtx = zapctx.ToContext(thunkCtx, logger)
-	thunkCtx = ioctx.StderrToContext(thunkCtx, rec.Stderr())
-
-	err = runHook(thunkCtx, hookThunk, bass.NewList(
-		bass.Bindings{
-			"event":   bass.String(eventName),
-			"payload": payloadScope,
-		}.Scope(),
-		(&bassgh.Client{
-			ExternalURL: c.externalURL,
-			DB:          c.DB,
-			GH:          ghClient,
-			Blobs:       c.Blobs,
-			Sender:      sender,
-			Repo:        repo,
-		}).Scope(),
-	))
-	if err != nil {
-		cli.WriteError(thunkCtx, err)
-	}
-
-	rec.Done(err)
-
-	if completeErr := runs.Record(ctx, c.DB, c.Blobs, run, progress, err == nil); completeErr != nil {
-		return fmt.Errorf("failed to complete: %w", completeErr)
-	}
-
-	return err
-}
-
-func runHook(ctx context.Context, hookThunk bass.Thunk, args bass.List) error {
-	logger := zapctx.FromContext(ctx)
-
-	// track thunk runs separately so we can log them later
-	ctx, runs := bass.TrackRuns(ctx)
-
-	module, err := bass.NewBass().Load(ctx, hookThunk)
-	if err != nil {
-		return fmt.Errorf("load project.bass: %w", err)
-	}
-
-	var comb bass.Combiner
-	if err := module.GetDecode("github-hook", &comb); err != nil {
-		return fmt.Errorf("get github-hook: %w", err)
-	}
-
-	logger.Info("calling github-hook")
-
-	call := comb.Call(ctx, args, module, bass.Identity)
-
-	if _, err := bass.Trampoline(ctx, call); err != nil {
-		return fmt.Errorf("hook: %w", err)
-	}
-
-	logger.Info("github-hook called; waiting on runs")
-
-	err = runs.Wait()
-	if err != nil {
-		logger.Warn("runs failed", zap.Error(err))
-	} else {
-		logger.Info("runs succeeded")
-	}
-
-	return err
 }
 
 func (c *Controller) withUserPool(ctx context.Context, user *github.User) (context.Context, *runtimes.Pool, error) {
@@ -322,5 +107,43 @@ func (c *Controller) withUserPool(ctx context.Context, user *github.User) (conte
 	}
 
 	return bass.WithRuntimePool(ctx, pool), pool, nil
+}
 
+func callHook(ctx context.Context, hookThunk bass.Thunk, hookBinding bass.Symbol, args bass.List) error {
+	logger := zapctx.FromContext(ctx).With(
+		zap.Stringer("thunk", hookThunk),
+		zap.Stringer("binding", hookBinding),
+	)
+
+	// track thunk runs separately so we can log them later
+	ctx, runs := bass.TrackRuns(ctx)
+
+	module, err := bass.NewBass().Load(ctx, hookThunk)
+	if err != nil {
+		return fmt.Errorf("load project.bass: %w", err)
+	}
+
+	var comb bass.Combiner
+	if err := module.GetDecode(hookBinding, &comb); err != nil {
+		return fmt.Errorf("get %s: %w", hookBinding, err)
+	}
+
+	logger.Info("calling hook")
+
+	call := comb.Call(ctx, args, module, bass.Identity)
+
+	if _, err := bass.Trampoline(ctx, call); err != nil {
+		return fmt.Errorf("hook: %w", err)
+	}
+
+	logger.Info("hook called; waiting on runs")
+
+	err = runs.Wait()
+	if err != nil {
+		logger.Warn("runs failed", zap.Error(err))
+	} else {
+		logger.Info("runs succeeded")
+	}
+
+	return err
 }
